Fix discarded suspended tabs showing as blank #256
Manifest V3 introduced a lot of async processing. As a result, the blank
suspended page was loading so fast, the tab state triggered from loading
to complete before state could be saved to session storage. It seems the
state management is unnecessary, and thus removed.
See comments in the commit.