Refactor gsFavicon following manifest V3 changes
The Manifest V3 conversion removed some of the favicon code due to fact
that V3 no longer allows DOM access from background scripts.  This
refactor brings it all back in.

The primary driver was the fact that YouTube favicons were failing to
load, but I suspect several of the icon-related issues will be improved
if not solved by this update.  #272, #268, #260