Update shortcuts page to fix the default action
This "fix" is needed because Manifest V3 changes the command name for
the default extension popup.  The original extension suppressed this
command entirely, but the migration to V3 re-enabled it due to the name
change.

This update sorta merges the two worlds.  The shortcut is displayed,
because Chrome's settings will display it too, but we give it a proper
i18n value.