Fix suspended page load race condition
The original migration to manifest V3 missed an "await" in tgs.  The
result was a race condition that would cause suspended tabs to
"quickInit" instead of fully loading on browser launch.

This might help with reported page load issues #276 #260

This commit also adjusts the automatic theme selection to use class
names instead of direct style visibility.  gsUtils was clobbering
the body classList, which is fixed here too.